refactor: 점수를 Score 객체가 관리하도록 리팩토링

+++ src/main/java/domain/participant/Dealer.java
@@ -11,6 +11,6 @@ public class Dealer extends Participant {
 
     @Override
     public boolean isReceivable() {
-        return score() <= THRESHOLD_SCORE;
+        return score().toInt() <= THRESHOLD_SCORE;
     }
 }
